Sprint code
This code will allow the player to sprint for a limited amount of time, to escape the mutant. The stamina bar has a visual change for the player to see, so that the player may decide what is the best decision to make, whether it is to keep sprinting, conserve energy, hide somewhere, etc.